♻️ Extract stream/vod handling to separate file
