+++ src/twitch/tally.ts
@@ -1,5 +1,5 @@
 import { ChatEvents, sendChatMessage } from './twitchChat.js'
-import { TwitchEvents } from './twitchEventSub.js'
+import { TwitchEvents } from './eventSub.js'
 
 type Message = { text: string; date: Date; userID: string }
 const recentMessages: Message[] = []
