+++ src/twitch/twitch.ts
@@ -2,7 +2,7 @@ import { initExpressApp } from '../express.js'
 import { initEmotes } from './emotes.js'
 import { createAuthAndApiClient } from './twitchApi.js'
 import { initTwitchChat } from './twitchChat.js'
-import { initTwitchEventSub } from './twitchEventSub.js'
+import { initTwitchEventSub } from './eventSub.js'
 import { initTwitchOAuthServer } from './twitchOAuth.js'
 
 export async function initTwitch() {
